create Circle class with fields and getter methods

+++ 04-design-from-scratch/src/main/java/it/unibo/shapes/api/Shape.java
@@ -1,7 +1,7 @@
 package it.unibo.shapes.api;
 
 public interface Shape {
-    float getArea();
-    
-    float getPerimeter();
+    double getArea();
+
+    double getPerimeter();
 }
